--- packages/ui/icons/.eslintrc.js
@@ -1,9 +1,0 @@
-/** @type {import("eslint").Linter.Config} */
-module.exports = {
-  root: true,
-  extends: ['favolink'],
-  parserOptions: {
-    project: true,
-    tsconfigRootDir: __dirname,
-  },
-};
