--- packages/ui/icons/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@favolink-ui/icons",
-  "version": "0.0.0",
-  "private": true,
-  "main": "src/index.ts",
-  "scripts": {
-    "lint": "eslint \"**/*.{js,jsx,cjs,mjs,ts,tsx,cts,mts}\""
-  },
-  "dependencies": {
-    "react": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.55"
-  }
-}
