+++ packages/ui/react/src/icons/bell.tsx
