+++ packages/ui/react/src/icons/cancel.tsx
