+++ packages/ui/react/src/icons/chevron-up.tsx
