+++ packages/ui/react/src/icons/close.tsx
