+++ packages/ui/react/src/icons/index.ts
@@ -1,5 +1,4 @@
 /* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export * from './icon';
 export * from './add';
 export * from './bell';
 export * from './cancel';
