+++ packages/ui/react/src/icons/trash.tsx
