+++ packages/ui/react/src/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export * from '@favolink-ui/icons';
+export * from './icons';
 export * from './hooks';
 export * from './styles';
 export * from './system';
